+++ packages/ssg-theme-daisyui/src/layouts/docs.tsx
@@ -10,6 +10,7 @@ import Header from '../components/Header.js';
 import Footer from '../components/Footer.js';
 import Sidebar from '../components/Sidebar.js';
 import TOC from '../components/TOC.js';
+import PrevNext from '../components/PrevNext.js';
 
 export default component<LayoutProps, unknown, LayoutSlots>(({ slots, props, signal }) => {
     const state = signal({
@@ -53,16 +54,17 @@ export default component<LayoutProps, unknown, LayoutSlots>(({ slots, props, sig
                         <div class="flex gap-8">
                             {/* Article content */}
                             <article class="flex-1 min-w-0 prose prose-lg max-w-none">
-                                {props.meta?.title && (
+                                {props.meta?.title && !props.meta?.titleFromContent && (
                                     <h1>{props.meta.title}</h1>
                                 )}
                                 {slots.default()}
+                                <PrevNext />
                             </article>
 
                             {/* Table of contents */}
                             <aside class="hidden xl:block w-64 shrink-0">
                                 <div class="sticky top-20">
-                                    <TOC />
+                                    <TOC meta={props.meta} />
                                 </div>
                             </aside>
                         </div>